refactor: 메뉴, 뷰 스테이트 형식 변경
